requirements.txt
asttokens==2.4.1
azure-core==1.30.2
azure-storage-blob==12.22.0
azure-storage-file-datalake==12.16.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
cryptography==43.0.1
debugpy==1.8.5
decorator==5.1.1
deltalake==0.19.1
duckdb==1.0.0
executing==2.1.0
idna==3.8
ipykernel==6.29.5
ipython==8.27.0
isodate==0.6.1
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==26.2.0
requests==2.32.3
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
